refactor(llm): extract LLM logic into llm folder
