+++ pdf_craft/__init__.py
@@ -1,4 +1,5 @@
 from .pdf import *
+from .llm import *
 from .markdown import *
 from .analyser import *
 from .epub import *
