+++ pdf_craft/analyser/__init__.py
@@ -1,3 +1,2 @@
-from .llm import LLM
 from .state_machine import analyse
 from .types import AnalysingStep, AnalysingProgressReport, AnalysingStepReport
