+++ pdf_craft/llm/node.py
